+++ lib.js
@@ -1,15 +1,16 @@
-import Err from './components/Err.vue'
-import Select from './components/Select.vue'
-import Modal from './components/Modal.vue'
-import Progress from './components/Progress.vue'
-import Input from './components/Input.vue'
-import Loading from './components/Loading.vue'
+import TablerError from './components/Err.vue'
+import TablerSelect from './components/Select.vue'
+import TablerModal from './components/Modal.vue'
+import TablerProgress from './components/Progress.vue'
+import TablerLoading from './components/Loading.vue'
+
+import TablerInput from './components/Input.vue'
 
 export {
-    Err,
-    Select,
-    Loading,
-    Modal,
-    Progress,
-    Input
+    TablerError,
+    TablerSelect,
+    TablerLoading,
+    TablerModal,
+    TablerProgress,
+    TablerInput
 }